+++ fixtures/gists-app/app/routes/render-error.js
@@ -1,0 +1,4 @@
+export default function RenderError() {
+  undefined();
+  return <div>You'll (probably) never get here!</div>;
+}